refactor(desktop): split agents page composition hub
